backend/requirements/base.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
django-countries==8.2.0
django-filter==25.2
django-jazzmin==3.0.0
django-js-asset==3.1.2
django-mptt==0.18.0
django-taggit==6.1.0
django-tinymce==5.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
idna==3.11
pillow==12.0.0
promise==2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
stripe==13.2.0
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.5.0
